requirements.txt
biopandas==0.4.1
einops==0.8.0
fair_esm==2.0.0
geatpy==2.7.0
imbalanced_learn==0.12.3
imblearn==0.0
ipython==8.12.3
keras_progbar==1.0.0
line_profiler==4.1.3
matplotlib==3.5.2
ml_collections==0.1.1
molecular_rectifier==1.0.1
networkx==2.7
numpy==1.19.5
openfold==0.0.1
pandarallel==1.6.5
pandas==1.2.2
pymoo==0.6.1.1
PyTDC==0.4.1
PyYAML==6.0.2
ray==2.9.1
rdkit==2023.9.5
Requests==2.32.3
scikit_learn==1.3.2
scipy==1.10.1
seaborn==0.11.2
setuptools==59.5.0
torch==1.10.2+cu113
torch_geometric==2.0.4
torch_scatter==2.0.9
tqdm==4.66.1
